Deploy project - Rename packages - Add circle config with test->build->deploy - Create amazon service to deploy
