+++ .gitignore
@@ -80,3 +80,4 @@ fabric.properties
 # .nfs files are created when an open file is removed but is still being accessed
 .nfs*
 
+/terraform_setup/terraform.tfvars
