+++ Dockerfile
@@ -1,0 +1,3 @@
+FROM golang:onbuild
+
+EXPOSE 8080:8080
