--- rest-api/main.go
@@ -1,11 +1,0 @@
-package api
-
-// main function
-func main() {
-    a := App{} 
-    // You need to set your Username and Password here
-    a.Initialize()
-
-    a.Run(":8080")
-
-}
