--- rest-api/model.go
@@ -1,8 +1,0 @@
-package api
-
-// model
-type Note struct {
-    ID        string   `json:"id,omitempty"`
-    Timestamp int      `json:"timestamp,omitempty"`
-    Text      string   `json:"text,omitempty"`
-}
