+++ terraform_setup/terraform.tfvars
@@ -1,0 +1,9 @@
+# Please fill in the values for all the variables here
+
+aws_access_key = ""
+aws_secret_key = ""
+aws_account_id = ""
+# e.g. us-east-1 (Please specify a region supported by Fargate launch type)
+aws_region     = ""
+# Prefix to be used in the naming of some of the created AWS resources. Example value: demo-webapp
+aws_resource_prefix = ""
